+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Converter for Bavaria, Germany
+
 ## [v0.21.0] - 2026-02-16
 
 - Update vecorel-cli